refactor: moved pagination to utils
